Mulesoft TemplateSpec Solution
