+++ Solutions/Mulesoft/Data Connectors/azuredeploy_Connector_MuleSoft_Cloudhub_API_AzureFunction.json
