+++ Solutions/Mulesoft/SolutionMetadata.json
@@ -1,0 +1,16 @@
+{
+	"publisherId": "azuresentinel",
+	"offerId": "azure-sentinel-solution-mulesoft",
+	"firstPublishDate": "2022-07-12",
+	"providers": ["MuleSoft"],
+	"categories": {
+		"domains" : ["DevOps","IT Operations"],
+		"verticals": []
+	},
+	"support": {
+	  "name": "Microsoft Corporation",
+	  "email": "support@microsoft.com",
+	  "tier": "Microsoft",
+	  "link": "https://support.microsoft.com"
+	}
+}
